requirements.txt
pandas
numpy
faker
kafka-python
pymongo
scikit-learn
mysql-connector-python
